refactor: move symbols to common
