+++ src/main/frontend/common/components/symbols.tsx
