+++ src/main/frontend/pipeline-console-view/pipeline-console/main/PipelineConsole.tsx
@@ -16,7 +16,7 @@ import { useLayoutPreferences } from "./providers/user-preference-provider.tsx";
 import ScrollToTopBottom from "./scroll-to-top-bottom.tsx";
 import SplitView from "./split-view.tsx";
 import StageView from "./StageView.tsx";
-import { CONSOLE, DOCUMENT, SETTINGS } from "./symbols.tsx";
+import { CONSOLE, DOCUMENT, SETTINGS } from "../../../common/components/symbols.tsx";
 
 export default function PipelineConsole() {
   const rootElement = document.getElementById("console-pipeline-root");
